refactor signup into smaller components
